functions from the module views.py have been moved to the module utils.py
